Moved decimation.py to multiscaling.
