--- src/skshapes/decimation/__init__.py
@@ -1,1 +1,0 @@
-from .decimation import Decimation
