+++ src/skshapes/multiscaling/__init__.py
@@ -1,3 +1,4 @@
 """Multiscale representation of shapes."""
 
+from .decimation import Decimation
 from .multiscale import Multiscale
